trace: Move azure tracer to exporters pkg
